refactor: 헤더 ui 분리
