+++ src/components/Header/Header.tsx
@@ -1,26 +1,26 @@
-import * as styles from './Header.style';
-import { useApiContext } from '../../contexts/ApiContext';
-import { CartItemResponse } from '../../types/response';
-import getCartItems from '../../api/getCartItems';
-import useErrorHandler from '../../hooks/useErrorHandler';
+import { css } from '@emotion/react';
+import { ReactNode } from 'react';
 
-function Header() {
-  const { data: cartItems, error: cartFetchError } = useApiContext<CartItemResponse>({
-    fetchFn: getCartItems,
-    key: 'getCartItems'
-  });
-
-  useErrorHandler(cartFetchError);
-
+export default function Header({ left, right }: { left: ReactNode; right: ReactNode }) {
   return (
-    <header css={styles.header}>
-      <p>SHOP</p>
-      <button css={styles.cartIcon}>
-        <img src="assets/cart.svg" alt="cart-icon" />
-        <span hidden={cartItems?.content.length === 0}>{cartItems?.content.length}</span>
-      </button>
+    <header css={headerCss}>
+      {left}
+      {right}
     </header>
   );
 }
 
-export default Header;
+const headerCss = css({
+  position: 'fixed',
+  top: '0%',
+  left: '50%',
+  transform: 'translate(-50%, 0)',
+  height: '64px',
+  width: '382px',
+  backgroundColor: 'black',
+  padding: '24px',
+  display: 'flex',
+  justifyContent: 'space-between',
+  alignItems: 'center',
+  zIndex: 100
+});
